Create an abstract superclass
